Extract `ShapeAndDTypeAllocator` from `Ones`; reparent its borrowers.
`Ones` carried no value semantics — its whole body is the explicit-shape
+ optional-dtype allocator machinery (the `SHAPE`/`DTYPE` parameter
slots, the float32 default, the "shape is mandatory" contract, the
constant-extraction helper), which is why nine unrelated generators
borrowed it via code-reuse-not-is-a `extends Ones` (wala/ML#514).

Move that machinery to a new abstract `ShapeAndDTypeAllocator` base and
reparent the borrowers (`Zeros`, `Input`, `OneHot`, `Uniform`,
`Poisson`, `Gamma`, `SparseTensor`, `ReadDataSets`) to it. `Ones` and
`NpOnes` (np.ones — a genuine kind of `ones`) keep their relationship.

Behavior-preserving: the base has identical methods to the old `Ones`,
and no code referenced `Ones.` directly. Full suite green (969 tests).
The `Eye extends SparseEye` inversion and the other `#514` clusters
follow in subsequent commits.
