+++ com.ibm.wala.cast.python.ml/source/com/ibm/wala/cast/python/ml/client/Ones.java
@@ -1,45 +1,17 @@
 package com.ibm.wala.cast.python.ml.client;
 
-import static com.ibm.wala.cast.python.ml.types.TensorFlowTypes.DType.FLOAT32;
-import static java.util.logging.Logger.getLogger;
-
-import com.ibm.wala.cast.python.ml.types.TensorFlowTypes.DType;
-import com.ibm.wala.cast.python.ml.types.TensorType.Dimension;
 import com.ibm.wala.ipa.callgraph.CGNode;
-import com.ibm.wala.ipa.callgraph.propagation.ConstantKey;
-import com.ibm.wala.ipa.callgraph.propagation.InstanceKey;
 import com.ibm.wala.ipa.callgraph.propagation.PointsToSetVariable;
-import com.ibm.wala.ipa.callgraph.propagation.PropagationCallGraphBuilder;
-import java.util.EnumSet;
-import java.util.List;
-import java.util.Locale;
-import java.util.Optional;
-import java.util.Set;
-import java.util.logging.Logger;
 
 /**
- * A generator for tensors created by the `ones()` function in TensorFlow.
+ * A generator for tensors created by the `ones()` function in TensorFlow. The explicit-shape +
+ * optional-dtype allocator machinery lives in {@link ShapeAndDTypeAllocator}; {@code tf.ones} adds
+ * no value semantics relevant to tensor-type inference.
  *
  * @see <a href="https://www.tensorflow.org/api_docs/python/tf/ones">TensorFlow ones() API</a>.
  * @author <a href="mailto:khatchad@hunter.cuny.edu">Raffi Khatchadourian</a>
  */
-public class Ones extends TensorGenerator {
-
-  private static final Logger LOGGER = getLogger(Ones.class.getName());
-
-  protected enum Parameters {
-    SHAPE,
-    DTYPE,
-    NAME;
-
-    public String getName() {
-      return name().toLowerCase(Locale.ROOT);
-    }
-
-    public int getIndex() {
-      return ordinal();
-    }
-  }
+public class Ones extends ShapeAndDTypeAllocator {
 
   public Ones(PointsToSetVariable source) {
     super(source);
@@ -47,52 +19,5 @@ public class Ones extends TensorGenerator {
 
   public Ones(CGNode node) {
     super(node);
-  }
-
-  @Override
-  protected Set<DType> getDefaultDTypes(PropagationCallGraphBuilder builder) {
-    LOGGER.fine(
-        "No dtype specified for source: " + source + ". Using default dtype of: " + FLOAT32 + " .");
-
-    // Use the default dtype of float32.
-    return EnumSet.of(FLOAT32);
-  }
-
-  @Override
-  protected Set<List<Dimension<?>>> getDefaultShapes(PropagationCallGraphBuilder builder) {
-    throw new UnsupportedOperationException("Shape is mandatory and must be provided explicitly.");
-  }
-
-  @Override
-  protected int getShapeParameterPosition() {
-    return Parameters.SHAPE.getIndex();
-  }
-
-  @Override
-  protected String getShapeParameterName() {
-    return Parameters.SHAPE.getName();
-  }
-
-  @Override
-  protected int getDTypeParameterPosition() {
-    return Parameters.DTYPE.getIndex();
-  }
-
-  @Override
-  protected String getDTypeParameterName() {
-    return Parameters.DTYPE.getName();
-  }
-
-  protected static Optional<Integer> getIntValueFromInstanceKey(InstanceKey instanceKey) {
-    if (instanceKey instanceof ConstantKey) {
-      ConstantKey<?> constantKey = (ConstantKey<?>) instanceKey;
-      Object value = constantKey.getValue();
-
-      if (value == null) return Optional.empty();
-      return Optional.of(((Number) value).intValue());
-    }
-
-    throw new IllegalArgumentException(
-        "Cannot get integer value from non-constant InstanceKey: " + instanceKey + ".");
   }
 }
